+++ examples/react-spa/frontend/src/lib/allauth.js
@@ -27,7 +27,7 @@ export const Flows = Object.freeze({
   MFA_AUTHENTICATE: 'mfa_authenticate',
   REAUTHENTICATE: 'reauthenticate',
   MFA_REAUTHENTICATE: 'mfa_reauthenticate',
-  MFA_WEBAUTHN_SIGNUP: 'mfa_webauthn_signup'
+  MFA_WEBAUTHN_SIGNUP: 'mfa_signup_webauthn'
 })
 
 export const URLs = Object.freeze({